+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add Overwatcher daily task to power cycle AG cameras before observations.
 * Check AG cameras before each tile.
 * Improve cancellation of the observe loop during startup.
+* Observe each standard for at least 60 seconds.
 
 
 ## 1.7.2 - March 2, 2025